Move the text search code to the :search into a separate package
Issue: #297
