+++ search/src/main/java/name/mlopatkin/andlogview/search/text/RegExpSearcher.java
@@ -13,28 +13,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package name.mlopatkin.andlogview.search;
+package name.mlopatkin.andlogview.search.text;
 
-import name.mlopatkin.andlogview.utils.MyStringUtils;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
 
-class IgnoreCaseSearcher implements HighlightStrategy, SearchStrategy {
-    private String textToSearch;
+class RegExpSearcher implements HighlightStrategy, SearchStrategy {
+    private final Pattern pattern;
 
-    public IgnoreCaseSearcher(String text) {
-        this.textToSearch = text;
+    public RegExpSearcher(String regex) throws PatternSyntaxException {
+        pattern = Pattern.compile(regex);
     }
 
     @Override
     public boolean test(String s) {
-        return MyStringUtils.indexOfIgnoreCase(s, textToSearch) >= 0;
+        return pattern.matcher(s).find();
     }
 
     @Override
     public void highlightOccurences(String text, TextHighlighter highlighter) {
-        int pos = MyStringUtils.indexOfIgnoreCase(text, textToSearch);
-        while (pos != MyStringUtils.NOT_FOUND && pos < text.length()) {
-            highlighter.highlightText(pos, pos + textToSearch.length());
-            pos = MyStringUtils.indexOfIgnoreCase(text, textToSearch, pos + textToSearch.length());
+        Matcher m = pattern.matcher(text);
+        int pos = 0;
+        while (m.find(pos)) {
+            highlighter.highlightText(m.start(), m.end());
+            pos = m.end();
         }
     }
 }
