+++ search/src/main/java/name/mlopatkin/andlogview/search/text/SearchStrategyFactory.java
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package name.mlopatkin.andlogview.search;
+package name.mlopatkin.andlogview.search.text;
+
+import name.mlopatkin.andlogview.search.RequestCompilationException;
 
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
