+++ search/src/main/java/name/mlopatkin/andlogview/search/text/TextHighlighter.java
@@ -13,13 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package name.mlopatkin.andlogview.search;
+package name.mlopatkin.andlogview.search.text;
 
 /**
- *
+ * An implementation of this interface can mark portions of text as highlighted.
  */
 public interface TextHighlighter {
+    /**
+     * Request highlighting of a text portion starting at char at {@code from} index (inclusive) and ending at char at
+     * {@code end} index (exclusive). The {@code end} index can be outside the text bounds if the highlighting spans to
+     * the end of text.
+     *
+     * @param from the index of the start of the highlighted portion of text
+     * @param to the first index after the end of the highlighted portion of text
+     */
     void highlightText(int from, int to);
-
-    void clearHighlight();
 }
