+++ src/name/mlopatkin/andlogview/search/RowSearchStrategyFactory.java
@@ -1,5 +1,23 @@
+/*
+ * Copyright 2013 Mikhail Lopatkin
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package name.mlopatkin.andlogview.search;
 
+import name.mlopatkin.andlogview.search.text.HighlightStrategy;
+import name.mlopatkin.andlogview.search.text.SearchStrategyFactory;
 import name.mlopatkin.andlogview.ui.logtable.Column;
 
 import com.google.common.base.CharMatcher;
