+++ src/name/mlopatkin/andlogview/search/ValueSearcher.java
@@ -17,6 +17,8 @@
 package name.mlopatkin.andlogview.search;
 
 import name.mlopatkin.andlogview.logmodel.LogRecord;
+import name.mlopatkin.andlogview.search.text.HighlightStrategy;
+import name.mlopatkin.andlogview.search.text.TextHighlighter;
 import name.mlopatkin.andlogview.ui.logtable.Column;
 
 class ValueSearcher implements RowSearchStrategy {
