+++ src/name/mlopatkin/andlogview/ui/logtable/HighlightCellRenderer.java
@@ -16,7 +16,7 @@
 package name.mlopatkin.andlogview.ui.logtable;
 
 import name.mlopatkin.andlogview.TooltipGenerator;
-import name.mlopatkin.andlogview.search.TextHighlighter;
+import name.mlopatkin.andlogview.search.text.TextHighlighter;
 import name.mlopatkin.andlogview.thirdparty.styledlabel.StyledLabel;
 import name.mlopatkin.andlogview.widgets.UiHelper;
 
@@ -122,8 +122,7 @@ public class HighlightCellRenderer extends StyledLabel implements TableCellRende
         document.setCharacterAttributes(from, to - from, highlighted, true);
     }
 
-    @Override
-    public void clearHighlight() {
+    private void clearHighlight() {
         document.setCharacterAttributes(0, document.getLength(), BASE_STYLE, true);
     }
 }
