+++ test/name/mlopatkin/andlogview/search/RowSearchersTest.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2013 Mikhail Lopatkin
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package name.mlopatkin.andlogview.search;
 
 import static org.junit.Assert.assertFalse;
@@ -7,6 +23,8 @@ import static org.mockito.Mockito.verifyNoInteractions;
 
 import name.mlopatkin.andlogview.logmodel.LogRecord;
 import name.mlopatkin.andlogview.logmodel.LogRecordUtils;
+import name.mlopatkin.andlogview.search.text.SearchStrategyFactory;
+import name.mlopatkin.andlogview.search.text.TextHighlighter;
 import name.mlopatkin.andlogview.ui.logtable.Column;
 
 import org.junit.Test;
